+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [12.6.9](///compare/v12.6.8...v12.6.9) (2021-05-29)
+
+### [12.6.8](///compare/v12.6.7...v12.6.8) (2021-05-28)
+
+
+### Bug Fixes
+
+* 27 May patch - Santa Ana addded 9454642
+
+### [12.6.7](///compare/v12.6.6...v12.6.7) (2021-05-27)
+
+### [12.6.6](///compare/v12.6.5...v12.6.6) (2021-05-27)
+
+
+### Bug Fixes
+
+* 27 May patch f2bec3b
+* 27 May patch - ships, ship blueprints 33777cd
+
+### [12.6.5](///compare/v12.6.4...v12.6.5) (2021-05-27)
+
+
+### Bug Fixes
+
+* 27 May patch f17b02c
+
+### [12.6.4](///compare/v12.6.3...v12.6.4) (2021-03-13)
+
+
+### Bug Fixes
+
+* **port investments:** February 2021 data 273633a
+
 ### [12.6.3](///compare/v12.6.2...v12.6.3) (2021-03-08)